+++ programs/drift/src/controller/amm.rs
@@ -167,6 +167,31 @@ pub fn update_spread_reserves(amm: &mut AMM) -> DriftResult {
 }
 
 pub fn update_spreads(amm: &mut AMM, reserve_price: u64) -> DriftResult<(u32, u32)> {
+    let max_offset = amm.get_max_reference_price_offset()?;
+
+    let reference_price_offset = if amm.curve_update_intensity > 0 {
+        let liquidity_ratio = amm_spread::calculate_inventory_liquidity_ratio(
+            amm.base_asset_amount_with_amm,
+            amm.base_asset_reserve,
+            amm.max_base_asset_reserve,
+            amm.min_base_asset_reserve,
+        )?;
+
+        amm_spread::calculate_reference_price_offset(
+            reserve_price,
+            amm.last_24h_avg_funding_rate,
+            liquidity_ratio,
+            amm.min_order_size,
+            amm.historical_oracle_data.last_oracle_price_twap_5min,
+            amm.last_mark_price_twap_5min,
+            amm.historical_oracle_data.last_oracle_price_twap,
+            amm.last_mark_price_twap,
+            max_offset,
+        )?
+    } else {
+        0
+    };
+
     let (long_spread, short_spread) = if amm.curve_update_intensity > 0 {
         amm_spread::calculate_spread(
             amm.base_spread,
@@ -196,6 +221,7 @@ pub fn update_spreads(amm: &mut AMM, reserve_price: u64) -> DriftResult<(u32, u3
 
     amm.long_spread = long_spread;
     amm.short_spread = short_spread;
+    amm.reference_price_offset = reference_price_offset;
 
     update_spread_reserves(amm)?;